Fix warnings in rtreetests

+++ src/query/tests/RTreeTests.cpp
@@ -422,7 +422,7 @@ BOOST_AUTO_TEST_CASE(ProjectionToVertex)
   BOOST_TEST(interpolation.second == 1.0);                           // Check the distance
   BOOST_TEST(interpolation.first.isInterpolation());
 
-  for (int i = 0; i < interpolation.first.getWeightedElements().size(); ++i) {
+  for (int i = 0; i < static_cast<int>(interpolation.first.getWeightedElements().size()); ++i) {
     BOOST_TEST(interpolation.first.getWeightedElements().at(i).vertexID == expectedIndices.at(i)); // Check index
     BOOST_TEST(interpolation.first.getWeightedElements().at(i).weight == expectedWeights.at(i));   // Check the weight
   }
@@ -466,7 +466,7 @@ BOOST_AUTO_TEST_CASE(ProjectionToTriangle)
   BOOST_TEST(interpolation.second == 0.0);                           // Check the distance
   BOOST_TEST(interpolation.first.isInterpolation());
 
-  for (int i = 0; i < interpolation.first.getWeightedElements().size(); ++i) {
+  for (int i = 0; i < static_cast<int>(interpolation.first.getWeightedElements().size()); ++i) {
     BOOST_TEST(interpolation.first.getWeightedElements().at(i).vertexID == expectedIndices.at(i)); // Check index
     BOOST_TEST(interpolation.first.getWeightedElements().at(i).weight == expectedWeights.at(i));   // Check the weight
   }
